Rename OpenTelemetry extension to observability
