+++ deno.json
@@ -16,7 +16,7 @@
     "./extensions/ext-db-sqlite",
     "./extensions/ext-parser-babel",
     "./extensions/ext-sandbox-shell-tools",
-    "./extensions/ext-tracing-opentelemetry",
+    "./extensions/ext-observability-opentelemetry",
     "./extensions/ext-content-mdx",
     "./extensions/ext-schema-zod"
   ],
@@ -85,7 +85,7 @@
     "./extensions/parser": "./src/extensions/parser/index.ts",
     "./extensions/sandbox": "./src/extensions/sandbox/index.ts",
     "./extensions/schema": "./src/extensions/schema/index.ts",
-    "./extensions/tracing": "./src/extensions/tracing/index.ts",
+    "./extensions/observability": "./src/extensions/observability/index.ts",
     "./testing": "./src/testing/index.ts",
     "./testing/assert": "./src/testing/assert.ts",
     "./testing/bdd": "./src/testing/bdd.ts",
@@ -180,7 +180,7 @@
     "veryfront/extensions/parser": "./src/extensions/parser/index.ts",
     "veryfront/extensions/sandbox": "./src/extensions/sandbox/index.ts",
     "veryfront/extensions/schema": "./src/extensions/schema/index.ts",
-    "veryfront/extensions/tracing": "./src/extensions/tracing/index.ts",
+    "veryfront/extensions/observability": "./src/extensions/observability/index.ts",
     "#veryfront": "./src/index.ts",
     "#veryfront/agent": "./src/agent/index.ts",
     "#veryfront/agent/react": "./src/agent/react/index.ts",
