+++ src/extensions/observability/tracing-exporter.ts
@@ -1,9 +1,9 @@
 /**
  * Contract interface for tracing/telemetry exporters.
  *
- * Default implementation: `@veryfront/ext-tracing-opentelemetry`
+ * Default implementation: `@veryfront/ext-observability-opentelemetry`
  *
- * @module extensions/tracing/tracing-exporter
+ * @module extensions/observability/tracing-exporter
  */
 
 /**
