+++ src/extensions/recommendations.ts
@@ -14,8 +14,8 @@ const recommendations = new Map<string, string>([
   ["ContentProcessor", "@veryfront/ext-content-mdx"],
   ["DocumentExtractor", "@veryfront/ext-document-kreuzberg"],
   ["AuthProvider", "@veryfront/ext-auth-jwt"],
-  ["TracingExporter", "@veryfront/ext-tracing-opentelemetry"],
-  ["NodeTelemetryProvider", "@veryfront/ext-tracing-opentelemetry"],
+  ["TracingExporter", "@veryfront/ext-observability-opentelemetry"],
+  ["NodeTelemetryProvider", "@veryfront/ext-observability-opentelemetry"],
   ["LLMProvider:openai", "@veryfront/ext-llm-openai"],
   ["LLMProvider:anthropic", "@veryfront/ext-llm-anthropic"],
   ["LLMProvider:google", "@veryfront/ext-llm-google"],
