+++ src/server/bootstrap.ts
@@ -13,7 +13,7 @@ import {
 } from "#veryfront/extensions/builtin-extensions.ts";
 import { MISSING_EXTENSION_ERROR } from "#veryfront/extensions/errors.ts";
 import { getRecommendation } from "#veryfront/extensions/recommendations.ts";
-import type { TracingExporter } from "#veryfront/extensions/tracing/tracing-exporter.ts";
+import type { TracingExporter } from "#veryfront/extensions/observability/tracing-exporter.ts";
 import {
   setGlobalActiveSpanAccessor,
   setGlobalMetricsAPI,
